multi_agent_debate/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.40
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
matplotlib>=3.7.0
pandas>=2.0.0
jupyter>=1.0.0
